DataConfiguration {esqlabsR}	R Documentation

DataConfiguration

Description

An object storing configuration of for observed data import

Super class

ospsuite::Printable -> DataConfiguration

Public fields

dataFolder

Path to the directory where the data file is located

dataFile

Name of the data excel file

compoundPropertiesFile

Name of the excel file with compound properties

dataSheets

Name of excel sheets to read

columnsToSplitBy

Column names by which the data will be split into groups

XValuesColumn

Column index for x values in observed data files

YValuesColumn

Column index for y values in observed data files

YErrorColumn

Column index for y error values in observed data files

Methods

Public methods

DataConfiguration$new()

DataConfiguration$print()

Inherited methods

Method new()

Initialize a new instance of the class

Usage

DataConfiguration$new(dataFolder, dataFile, compoundPropertiesFile, dataSheets)

Arguments

dataFolder

Path to the directory where the data file is located

dataFile

Name of the data excel file

compoundPropertiesFile

Name of the excel file with compound properties

dataSheets

Name of excel sheets to read

Returns

A new DataConfiguration object.

Method print()

Print the object to the console

Usage

DataConfiguration$print(...)

Arguments

...

Rest arguments.

[Package esqlabsR version 1.2.3 Index]