OR {esqlabsR}	R Documentation

Convenience function to avoid testing for null. It returns the first object if it is not null otherwise the second object

Description

Convenience function to avoid testing for null. It returns the first object if it is not null otherwise the second object

Usage

lhs %||% rhs

Arguments

lhs
Object that will be returned if not NULL

rhs
Object that will be returned if lhs is NULL. It maybe well be NULL

Value

The first parameter if it is not NULL otherwise the second parameter

[Package esqlabsR version 1.2.3 Index]