PlotTypes {esqlabsR}	R Documentation

Possible entries for the plotType field of a DataMapping object

Description

Possible entries for the plotType field of a DataMapping object

Usage

PlotTypes

Format

An object of class list of length 4.

Details

"IndividualProfile" - simulated results are plotted as time-values series with points connected by lines, with each individual results potted separately "PopulationQuantiles" - simulated results for a population are aggregated as median, 95th an 5th percentiles, with median plotted as a line an upper/lower percentiles plotted as shaded areas "PredictedVsObserved" - predicted-versus-observed goodness of fit plot

[Package esqlabsR version 1.2.3 Index]