Plotable {esqlabsR}	R Documentation

Plotable

Description

An object holding plotting information

Super class

ospsuite::Printable -> Plotable

Public fields

label

A string that is used as a label (e.g. in the legend) for the data set that will be plotted

color

Color that will be used when plotting the output. Represented as a array of three HSV values.

xFactor

Numeric value the x-values will be multiplied by.

yFactor

Numeric value the y-values will be multiplied by.

xOffset

A value that is added to all x-values when plotting

yOffset

A value that is added to all y-values when plotting

type

"p" for points, "l" for line, or "pl" for both

pch

Either an integer specifying a symbol or a single character to be used as the default in plotting points. See par for more information.

lty

The line type. See par for more information.

Methods

Public methods

Plotable$new()

Plotable$print()

Inherited methods

Method new()

Initialize a new instance of the class

Usage

Plotable$new(label)

Arguments

label

A string that is used as a label (e.g. in the legend) for the data set

Returns

A new Plotable object.

Method print()

Print the object to the console

Usage

Plotable$print(...)

Arguments

...

Rest arguments.

See Also

col2hsv

[Package esqlabsR version 1.2.3 Index]