applyIndividualParameters {esqlabsR}	R Documentation

Apply an individual to the simulation. For human species, only parameters that do not override formulas are applied. For other species, all parameters returned by createIndividual are applied.

Description

Apply an individual to the simulation. For human species, only parameters that do not override formulas are applied. For other species, all parameters returned by createIndividual are applied.

Usage

applyIndividualParameters(individualCharacteristics, simulation)

Arguments

individualCharacteristics
IndividualCharacteristics describing an individual. Optional

simulation
Simulation loaded form the PKML file

Examples

#'
## Not run: 
simulation <- loadSimulation(filePath = modelPath)
humanIndividualCharacteristics <- createIndividualCharacteristics(
  species = Species$Human, population = HumanPopulation$European_ICRP_2002,
  gender = Gender$Male, weight = 70
)
applyIndividualParameters(humanIndividualCharacteristics, simulation)

## End(Not run)

[Package esqlabsR version 1.2.3 Index]