Calculate the root mean square error for groupings in dataMappings

Calculate the root mean square error for groupings in dataMappings

Usage

calculateRMSE(dataMappingList, timeDiffThreshold = 10)

Arguments

dataMappingList

A DataMapping or a list of DataMapping objects.

timeDiffThreshold

Allowed difference between observed and simulated time values in minutes. Default is 10. If for a certain observed point no simulated time point exists within the defined threshold, the value is not considered.

Value

Total error for all groups across all provided data mappings.

Details

The error is calculated for each group separately and added up. For each group, the error is defined as the root of the sum of the squared residuals between each simulated result and observed data.