Compare all parameters of two simulations

The method compares all parameters of the two simulations and return a named list with lists In1NotIn2 (paths of parameters present in simulation1 but not in simulation2), In2NotIn1 (paths of parameters present in simulation2 but not in simulation1), and Different a list of paths of parameters that differ between simulation1 and simulation1. Two parameters are considered different if their formulas or values differ

compareSimulationParameters(simulation1, simulation2)

Arguments

simulation1
First Simulation to compare

simulation2
Second Simulation to compare

Value

Named list with entries In1NotIn2, In2NotIn1, and Different, holding the paths that are present in the first but not in the second simulation, present in the second but not in the first simulation, and present in both simulations but with different formulas and/or values, respectively.

See also

Examples

if (FALSE) {
humanSim <- loadSimulation(file.path(modelFolder, "DefaultHuman.pkml"))
ratSim <- loadSimulation(file.path(modelFolder, "DefaultRat.pkml"))
diffParams <- compareSimulationParameters(humanSim, ratSim)
}