.figureAddLegend {esqlabsR}	R Documentation

Add legend to a plot

Description

Add legend to a plot

Usage

.figureAddLegend(x, legend, col, pch, lty, ...)

Arguments

x
the x and y co-ordinates to be used to position the legend. They can be specified by keyword or in any way which is accepted by xy.coords: See ‘Details’.

legend
a character or expression vector of length ≥ 1 to appear in the legend. Other objects will be coerced by as.graphicsAnnot.

col
the color of points or lines appearing in the legend.

pch
the plotting symbols appearing in the legend, as numeric vector or a vector of 1-character strings (see points). Unlike points, this can all be specified as a single multi-character string. Must be specified for symbol drawing.

lty
the line types and widths for lines appearing in the legend. One of these two must be specified for line drawing.

...
Additional arguments to be passed to graphics::legend.

[Package esqlabsR version 1.2.3 Index]