Export steady-state to excel

Export steady-state to excel

Usage

exportSteadyStateToXLS(
  simulation,
  quantitiesPaths = NULL,
  resultsXLSPath = "",
  steadyStateTime = 1000,
  ignoreIfFormula = TRUE,
  stopIfNotFound = TRUE,
  lowerThreshold = 1e-15,
  simulationRunOptions = NULL
)

Arguments

simulation

A Simulation object that will be updated with the steady state

quantitiesPaths

List of quantity paths (molecules and/or parameters) for which the steady-state will be simulated. If NULL (default), all molecules and state variable parameters are considered. The same list is applied for all simulations.

resultsXLSPath

Path to the xls-file where the results will be written to. If the file does not exist, a new file is created. If no path is provided, the file will be created in the same directory where the model file is located. The name of the file will be <SimulationFileName>_SS.

steadyStateTime

Simulation time (minutes). Must be long enough for system to reach a steady-state. 1000 by default

ignoreIfFormula

If TRUE (default), species and parameters with initial values defined by a formula are not included.

stopIfNotFound

Boolean. If TRUE (default), an error is thrown when results for certain species were not generated. This may happen when due to numerical problems some values cannot be calculated, though the whole simulation converges. Setting this argument to FALSE allows to ignore such errors. Check the outputs for empty values when using this option.

lowerThreshold

Numerical value (in default unit of the output). Any steady-state values below this value are considered as numerical noise and replaced by 0. If lowerThreshold is NULL, no cut-off is applied. Default value is 1e-15.

simulationRunOptions

Optional instance of a SimulationRunOptions used during the simulation run.

Details

Simulates a given model to its steady-state and creates an Excel-file with the end values of molecules amounts in all containers and parameter values that have a right-hand-side (state variable parameters).