foldChangeFunction {esqlabsR}	R Documentation

Function returning the fold difference between x and x_0

Description

Function returning the fold difference between x and x_0

Usage

foldChangeFunction(x, x_0, alpha = 1)

Arguments

x
Value of the effector.

x_0
Basal value of the effector.

alpha
Steepness of the function. Default is 1.

Value

(x / x_0)^alpha

[Package esqlabsR version 1.2.3 Index]