Get names of all observed data in a Project

Returns the names of all DataSets that would be returned by loadObservedData(). On first call this loads the data to discover names; subsequent calls return cached names until a mutation invalidates the cache.

Usage

getObservedDataNames(project)

Arguments

project

A Project object (see loadProject()).

Value

A character vector of DataSet names.

See also