getSteadyState {esqlabsR}	R Documentation

Get the steady-state values of species and state variable parameters.

Description

Get the steady-state values of species and state variable parameters.

Usage

getSteadyState(
  quantitiesPaths = NULL,
  simulations,
  steadyStateTime,
  ignoreIfFormula = TRUE,
  stopIfNotFound = TRUE,
  lowerThreshold = 1e-15
)

Arguments

quantitiesPaths
List of quantity paths (molecules and/or parameters) for which the steady-state is to be simulated. If NULL (default), all molecules and state variable parameters are considered. The same list is applied for all simulations.

simulations
Simulation object or a list of Simulation objects

steadyStateTime
Simulation time (minutes). Must be long enough for system to reach a steady-state. 1000 by default

ignoreIfFormula
If TRUE (default), species and parameters with initial values defined by a formula are not included.

stopIfNotFound
Boolean. If TRUE (default), an error is thrown when results for certain species were not generated. This may happen when due to numerical problems some values cannot be calculated, though the whole simulation converges. Setting this argument to FALSE allows to ignore such errors. Check the outputs for empty values when using this option.

lowerThreshold
Numerical value (in µmol). Any steady-state values below this value are considered as numerical noise and replaced by 0. If lowerThreshold is NULL, no cut-off is applied. Default value is 1e-15.

Details

The steady-state is considered to be the last value of the simulation with sufficiently long simulation time, i.e., where the rates of the processes do not (significantly) change.

Value

A named list, where the names are the IDs of the simulations and the entries are lists containing paths and their values at the end of the simulation.

[Package esqlabsR version 1.2.3 Index]