hillFunction {esqlabsR}	R Documentation

Hill equation, transforming a value of an effector (e.g. concentration) to a saturable function value.

Description

Hill equation, transforming a value of an effector (e.g. concentration) to a saturable function value.

Usage

hillFunction(x, Vmax, Km, alpha = 1)

Arguments

x
Value of the effector.

Vmax
Maximal value of the function.

Km
Value of the effector at which the value of the function is half-maximal.

alpha
Hill coefficient defining the steepness of the function. Default is 1.

Value

Vmax * x^alpha / (x^alpha + Km^alpha)

[Package esqlabsR version 1.2.3 Index]