initializeSimulation {esqlabsR}	R Documentation

Load a simulation and apply a set of parameters.

Description

Helper method that combines a set of common steps performed before running a simulation. This method applies individual parameters data set and additional user-defined parameters to the simulation and runs the simulation to its steady-state and applies the steady-state as new initial conditions.

Usage

initializeSimulation(
  simulation,
  individualCharacteristics = NULL,
  additionalParams = NULL,
  simulateSteadyState = FALSE,
  steadyStateTime = 1000,
  ignoreIfFormula = TRUE,
  stopIfParameterNotFound = TRUE
)

Arguments

simulation
Simulation loaded from a PKML file

individualCharacteristics
Optional IndividualCharacteristics describing an individual.

additionalParams
Optional named list with lists 'paths', 'values', and 'units'.

simulateSteadyState
Logical. If TRUE, the model is simulated for steadyStateTime minutes after applying parameter values defined in individualCharacteristics and additionalParams, and the end results of the simulation are applied as initial conditions for all molecules. Default is FALSE.

steadyStateTime
Simulation time (minutes) for the steady-state simulation. Must be long enough for system to reach a steady-state 1000 by default.

ignoreIfFormula
If TRUE (default), species and parameters with initial values defined by a formula are not included in the steady-state simulation

stopIfParameterNotFound
Logical. If TRUE (default), an error is thrown if any of the additionalParams does not exist. If FALSE, non-existent parameters are ignored.

Examples

## Not run: 
simulation <- loadSimulation(filePath = modelPath)
humanIndividualCharacteristics <- createIndividualCharacteristics(
  species = Species$Human, population = HumanPopulation$European_ICRP_2002,
  gender = Gender$Male, weight = 70
)
userParams <- readParametersFromXLS(parameterXLSPath)
initializeSimulation(simulation, humanIndividualCharacteristics, userParams)
simulationResults <- runSimulation(simulation = simulation)

## End(Not run)

[Package esqlabsR version 1.2.3 Index]