isTableFormulasEqual {esqlabsR}	R Documentation

Check if two table formulas are equal.

Description

Table formulas are equal if the number of points is equal and all x-y value pairs are equal between the two formulas

Usage

isTableFormulasEqual(formula1, formula2)

Arguments

formula1
First formula to compare

formula2
Second formula to compare

Value

TRUE if the table formulas are equal, FALSE otherwise

[Package esqlabsR version 1.2.3 Index]