openOuptutDevice {esqlabsR}	R Documentation

Open an output device.

Description

Open an output device.

Usage

openOuptutDevice(plotConfiguration, width, height)

Arguments

plotConfiguration
An object of type PlotConfiguration

width
Width of the output figure

height
Height of the output figure

Details

If the output of the plot is directed to a file, open the connection. A list of supported outputs is provided in GraphicsDevices-enum. If the provided output defined in PlotConfiguration$outputDevice is not supported or the value is NULL, output is directed to the default plot frame.

[Package esqlabsR version 1.2.3 Index]