plotErrorBars {esqlabsR}	R Documentation

A function to add error bars on the chart.

Description

Taken from https://www.r-graph-gallery.com/4-barplot-with-error-bar.html

Usage

plotErrorBars(
  x,
  y,
  upper,
  lower = upper,
  length = par()$cin[[1]]/2,
  axis = "y",
  ...
)

Arguments

x
Numerical array of x-values

y
Numerical array of y-values

upper
Numerical array of upper y-error values

lower
Numerical array of lower y-error values. Optional. If not specified, same values as for upper are used

length
Numerical value specifying the width of the error bars. Optional. Default is 0.1.

axis
Dimension to which the error bars are added. If "y" (default), vertical error bars are drawn. If "x", horizontal error bars are drawn

...
Graphical parameters (see par())

[Package esqlabsR version 1.2.3 Index]