plotPredictedVsObserved {esqlabsR}	R Documentation

Plot a predicted-versus-observed goodness of fit plot

Description

Plot a predicted-versus-observed goodness of fit plot

Usage

plotPredictedVsObserved(
  dataMapping,
  foldDistance = 2,
  timeDiffThreshold = 10,
  ...
)

Arguments

dataMapping
THe DataMapping object for which the goodness-of-fit plot is to be drawn. For each group within the dataMapping, simulated and observed values are compared.

foldDistance
Numerical value for the fold-distance lines to be drawn. Default is 2.

timeDiffThreshold
Allowed difference between observed and simulated time values in minutes. Default is 10. If for a certain observed point no simulated time point exists within the defined threshold, the value is not considered.

...
Any parameter that can be interpreted by the default plot() function

Details

Observed data points are drawn on the x, simulated values on the y axis.

[Package esqlabsR version 1.2.3 Index]