plotXYData {esqlabsR}	R Documentation

Plot XYData

Description

Draw XYData on top of an existing plot using the points method.

Usage

plotXYData(xySeries, xUnit = NULL, yUnit = NULL, ...)

Arguments

xySeries
An XYData object to be plotted

xUnit
Target unit of x-axis.

yUnit
Target unit of y-axis. If TRUE,

...
Any parameter that can be interpreted by the default plot() function

[Package esqlabsR version 1.2.3 Index]