plotXYDataAggregated {esqlabsR}	R Documentation

Plot XYData quantiles

Description

Draw XYData on top of an existing plot using the points method.

Usage

plotXYDataAggregated(
  xySeries,
  xUnit = NULL,
  yUnit = NULL,
  quantiles = c(0.05, 0.5, 0.95),
  ...
)

Arguments

xySeries
An XYData object to be plotted

xUnit
Target unit of x-axis.

yUnit
Target unit of y-axis.

quantiles
A numerical vector with quantile values. Default is c(0.05, 0.5, 0.95)

...
Any parameter that can be interpreted by the default plot() function

[Package esqlabsR version 1.2.3 Index]