Read XLSX files using readxl::read_excel with suppressed warnings

Read XLSX files using readxl::read_excel with suppressed warnings

Usage

readExcel(path, sheet = NULL, ...)

Arguments

path

Full path of an XLS/XLSX file

sheet

Name or number of the sheet. If NULL (default), the first sheet of the file is used.

...

Any other parameters that can be passed to readxl::read_excel

Value

A tibble with the contents of the excel sheet