readOSPSTimeValues {esqlabsR}	R Documentation

Read time-values data from excel file

Description

Read time-values data from excel file

Usage

readOSPSTimeValues(dataConfiguration)

Arguments

dataConfiguration
An object of DataConfiguration

Details

The methods reads time-values data from the properly defined excel sheet and creates XYData objects according to the groupings. Each sheet in DataConfiguration$sheets is split according to columns listed in DataConfiguration$columnsToSplitBy. The output structure is a nested list with levels corresponding to the groupings.

Value

A (nested) list of XYData objects

[Package esqlabsR version 1.2.3 Index]