Carry out and visualize sensitivity analysis (with OSPSuite)

Carry out and visualize sensitivity analysis (with OSPSuite)

Usage

sensitivityCalculation(
  simulation,
  outputPaths,
  parameterPaths,
  variationRange = c(seq(0.1, 1, by = 0.1), seq(2, 10, by = 1)),
  pkParameters = c("C_max", "t_max", "AUC_inf"),
  pkDataFilePath = NULL,
  simulationRunOptions = NULL
)

Arguments

simulation

An object of type Simulation.

outputPaths

Path (or a vector of paths) to the output(s) for which the sensitivity will be analyzed.

parameterPaths

A single or a vector of the parameter path(s) to be varied.

variationRange

Optional numeric vector defining the scaling of the parameters. The same variation range is applied to all specified parameters. If not specified, the following vector will be used: c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10).

pkParameters

A vector of names of PK parameters for which the sensitivities will be calculated. For a full set of available standard PK parameters, run names(ospsuite::StandardPKParameter). By default, only the following parameters will be considered: "C_max", "t_max", "AUC_inf". If NULL, all available PK-parameters will be calculated. You can also specify custom PK parameters.

pkDataFilePath

Path to excel file in which PK-parameter data should be saved. If a file already exists, it will be overwritten. Default is NULL, meaning the data will not be saved to a spreadsheet.

simulationRunOptions

Optional instance of a SimulationRunOptions used during the simulation run

Value

A list containing following objects:

SimulationResults

specified output paths

specified parameter paths

A data frame of PK parameters

See also

Examples

# \dontrun{
simPath <- system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
simulation <- loadSimulation(simPath)
outputPaths <- "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)"
parameterPaths <- c(
  "Aciclovir|Lipophilicity",
  "Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose",
  "Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction"
)

# extract the results into a list of dataframes
sensitivityCalculation(
  simulation = simulation,
  outputPaths = outputPaths,
  parameterPaths = parameterPaths
)
#> $simulationResults
#> $simulationResults$`Aciclovir|Lipophilicity`
#> $simulationResults$`Aciclovir|Lipophilicity`$`0.1`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`0.2`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`0.3`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`0.4`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`0.5`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`0.6`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`0.7`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`0.8`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`0.9`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`1`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`2`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`3`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`4`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`5`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`6`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`7`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`8`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`9`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Aciclovir|Lipophilicity`$`10`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`0.1`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`0.2`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`0.3`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`0.4`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`0.5`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`0.6`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`0.7`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`0.8`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`0.9`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`1`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`2`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`3`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`4`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`5`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`6`
#> SimulationResults: 
#>    Number of individuals: 0 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`7`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`8`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`9`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose`$`10`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`0.1`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`0.2`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`0.3`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`0.4`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`0.5`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`0.6`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`0.7`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`0.8`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`0.9`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`1`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`2`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`3`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`4`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`5`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`6`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`7`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`8`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`9`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> $simulationResults$`Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction`$`10`
#> SimulationResults: 
#>    Number of individuals: 1 
#> 
#> 
#> 
#> $outputPaths
#> [1] "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)"
#> 
#> $parameterPaths
#> [1] "Aciclovir|Lipophilicity"                                                            
#> [2] "Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose"                  
#> [3] "Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction"
#> 
#> $pkData
#> # A tibble: 168 x 9
#>    OutputPath           ParameterPath ParameterFactor ParameterValue PKParameter
#>    <chr>                <chr>                   <dbl>          <dbl> <chr>      
#>  1 Organism|Peripheral~ Aciclovir|Li~             0.1        -0.0097 AUC_inf    
#>  2 Organism|Peripheral~ Aciclovir|Li~             0.2        -0.0194 AUC_inf    
#>  3 Organism|Peripheral~ Aciclovir|Li~             0.3        -0.0291 AUC_inf    
#>  4 Organism|Peripheral~ Aciclovir|Li~             0.4        -0.0388 AUC_inf    
#>  5 Organism|Peripheral~ Aciclovir|Li~             0.5        -0.0485 AUC_inf    
#>  6 Organism|Peripheral~ Aciclovir|Li~             0.6        -0.0582 AUC_inf    
#>  7 Organism|Peripheral~ Aciclovir|Li~             0.7        -0.0679 AUC_inf    
#>  8 Organism|Peripheral~ Aciclovir|Li~             0.8        -0.0776 AUC_inf    
#>  9 Organism|Peripheral~ Aciclovir|Li~             0.9        -0.0873 AUC_inf    
#> 10 Organism|Peripheral~ Aciclovir|Li~             1          -0.097  AUC_inf    
#> # ... with 158 more rows, and 4 more variables: PKParameterValue <dbl>,
#> #   Unit <chr>, PercentChangePK <dbl>, SensitivityPKParameter <dbl>
#> 
#> attr(,"class")
#> [1] "SensitivityCalculation" "list"                  
# }