stringToNum {esqlabsR}	R Documentation

Convert string to numeric

Description

Convert string to numeric

Usage

stringToNum(string)

Arguments

string
A string or a list of strings to be converted to numeric values

Details

Tries to convert each string to a numeric with as.numeric If any conversion fails and returns an NA, the value is tested for being a LLOQ-value, i.e., of a form "<2". If this is a case, the value is substituted by 0. In any other case where the string cannot be converted to a numeric, an NA is returned.

Value

A numeric value or a list of numeric values

[Package esqlabsR version 1.2.3 Index]