writeIndividualToXLS {esqlabsR}	R Documentation

Create a parameter set describing an individual and write it to the Excel file

Description

Create a parameter set describing an individual and write it to the Excel file

Usage

writeIndividualToXLS(individualCharacteristics, outputXLSPath)

Arguments

individualCharacteristics
An IndividualCharacteristics object describing the individual. See createIndividualCharacterstics for more information.

outputXLSPath
Path to the Excel file the parameter set will be written to

See Also

createIndividualCharacteristics crateIndividual

Examples

## Not run: 
simulation <- loadSimulation(pathToPKML)
humanIndividualCharacteristics <- createIndividualCharacteristics(
  species = Species$Human, population = HumanPopulation$European_ICRP_2002,
  gender = Gender$Male, weight = 70
)
writeIndividualToXLS(humanIndividualCharacteristics, pathToExcelFile)

## End(Not run)

[Package esqlabsR version 1.2.3 Index]