DataMappingConfiguration

An object storing configuration of a data mapping

Super class

ospsuite::Printable -> DataMappingConfiguration

Active bindings

xFactors

A named list listing numerical values that x-values are multiplied by. Keys are names of data sets, values are numerical factors.

yFactors

A named list listing numerical values that y-values are multiplied by. Keys are names of data sets, values are numerical factors.

xOffsets

A named list listing numerical values that will be added to the x-values. Keys are names of data sets, values are numerical values

yOffsets

A named list listing numerical values that will be added to the y-values. Keys are names of data sets, values are numerical values

lineTypes

A named list listing line types that will be used to plot a certain data set. Keys are names of data sets, values are numerical values recognized by the lty argument of the plot() function

colors

A named list listing color values that will be used to plot a certain data set. Keys are names of data sets, values are values recognized by the col argument of the plot() function

Methods

Public methods

DataMappingConfiguration$new()

DataMappingConfiguration$finalize()

DataMappingConfiguration$setXFactors()

DataMappingConfiguration$setYFactors()

DataMappingConfiguration$setXOffsets()

DataMappingConfiguration$setYOffsets()

DataMappingConfiguration$setColors()

DataMappingConfiguration$setLineTypes()

DataMappingConfiguration$print()

Inherited methods

Method new()

Initialize a new instance of the class

Usage

DataMappingConfiguration$new()

Returns

A new DataMappingConfiguration object.

Method finalize()

Clean up upon object removal

Usage

DataMappingConfiguration$finalize()

Method setXFactors()

Set the values of x-factors. An x-factor will be multiplied by the x-values of the data set.

Usage

DataMappingConfiguration$setXFactors(labels, xFactors)

Arguments

labels

A string or a list of strings representing the names of the data sets

xFactors

A numerical value or a list of numerical values representing the x-factors. Both lists must be of same length and store the entries in the same order.

Method setYFactors()

Set the values of y-factors. A y-factor will be multiplied by the y-values of the data set.

Usage

DataMappingConfiguration$setYFactors(labels, yFactors)

Arguments

labels

A string or a list of strings representing the names of the data sets

yFactors

A numerical value or a list of numerical values representing the y-factors. Both lists must be of same length and store the entries in the same order.

Method setXOffsets()

Set the values of x-offsets. An x-offset will be added to the the x-values of the data set.

Usage

DataMappingConfiguration$setXOffsets(labels, xOffsets)

Arguments

labels

A string or a list of strings representing the names of the data sets

xOffsets

A numerical value or a list of numerical values representing the x-offsets. Both lists must be of same length and store the entries in the same order.

Method setYOffsets()

Set the values of y-offsets. A y-offset will be added to the the y-values of the data set.

Usage

DataMappingConfiguration$setYOffsets(labels, yOffsets)

Arguments

labels

A string or a list of strings representing the names of the data sets

yOffsets

A numerical value or a list of numerical values representing the y-offsets. Both lists must be of same length and store the entries in the same order.

Method setColors()

Set the colors that will be used to plot a certain data set.

Usage

DataMappingConfiguration$setColors(labels, colors)

Arguments

labels

A string or a list of strings representing the names of the data sets

colors

Values representing a color recognized by the col argument of the plot() function

Method setLineTypes()

Set the line types that will be used to plot a certain data set.

Usage

DataMappingConfiguration$setLineTypes(labels, lineTypes)

Arguments

labels

A string or a list of strings representing the names of the data sets

lineTypes

Values recognized by the col argument of the plot() function

Method print()

Print the object to the console

Usage

DataMappingConfiguration$print(...)

Arguments

...

Rest arguments.