ExportConfiguration

R6 class defining properties for saving a ggplot object

Value

The file name of the exported plot

Super class

tlf::ExportConfiguration -> ExportConfiguration

Public fields

name

character defining the name of the file to be saved (without extension)

path

Path of the directory to save plot to: path and filename are combined to create the fully qualified file name. Defaults to the working directory.

format

character defining the format of the file to be saved

width

numeric values defining the width in units of the plot dimensions after saving

height

numeric values defining the height in units of the plot dimensions after saving. Only used if heightPerRow is NULL.

units

character defining the unit of the saving dimension

dpi

(dots per inch) numeric value defining plot resolution

Active bindings

name

character defining the name of the file to be saved (without extension)

path

Path of the directory to save plot to: path and filename are combined to create the fully qualified file name. Defaults to the working directory.

format

character defining the format of the file to be saved

width

numeric values defining the width in units of the plot dimensions after saving

height

numeric values defining the height in units of the plot dimensions after saving. Only used if heightPerRow is NULL.

units

character defining the unit of the saving dimension

dpi

(dots per inch) numeric value defining plot resolution

heightPerRow

The height of the plot dimensions for a row in a multi panel plot. The final height of the figure will be 'heightPerRow' times the number of rows. If NULL (default), value used in height is used. If not NULL, this value always overrides the height property.

Methods

Public methods

ExportConfiguration$savePlot()

ExportConfiguration$clone()

Inherited methods

tlf::ExportConfiguration$convertPixels()

tlf::ExportConfiguration$getFileName()

tlf::ExportConfiguration$initialize()

tlf::ExportConfiguration$print()

Method savePlot()

Save/Export a plot

Usage

ExportConfiguration$savePlot(plotObject, fileName = NULL)

Arguments

plotObject

A ggplot object

fileName

character file name of the exported plot

Method clone()

The objects of this class are cloneable with this method.

Usage

ExportConfiguration$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.