PlotConfiguration {esqlabsR}	R Documentation

PlotConfiguration

Description

An object storing configuration of a plot

Super class

ospsuite::Printable -> PlotConfiguration

Public fields

outputDevice

Output target of the plot. If NULL (default), the figure is created in the default "plot" output. Other values indicate output into a file. A list of supported outputs is provided in GraphicsDevices-enum.

outputName

A string used as the name of the .png file and as the title of the plot

outputFolder

Path to the directory where the outputs should be stored.

width

Width of the resulting plot in cm. If NULL (default), the width is automatically calculated using the value esqlabsEnv$widthPerPlotMapping

height

Height of the resulting plot in cm. If NULL (default), the height is automatically calculated using the value esqlabsEnv$heightPerPlotMapping

nrOfCols

Number of column in a multi-panel plot. If NULL (default), the number is calculated automatically to fit all panels while trying to keep the number of columns and rows equal. If nrOfCols is specified, the number of rows calculated to fit all the panels.

res

Resolution of the .png output in dpi. Default is 600.

pointsize

Size of the plotted text. Default is 8.

addTitle

Boolean flag if the title should be added. If TRUE, outputName is added a title.

Methods

Public methods

PlotConfiguration$new()

PlotConfiguration$print()

Inherited methods

Method new()

Initialize a new instance of the class

Usage

PlotConfiguration$new()

Returns

A new PlotConfiguration object.

Method print()

Print the object to the console

Usage

PlotConfiguration$print(...)

Arguments

...

Rest arguments.

[Package esqlabsR version 1.2.3 Index]