ProjectConfiguration

An object storing configuration used project-wide

Super class

ospsuite.utils::Printable -> ProjectConfiguration

Active bindings

projectConfigurationFilePath

Path to the file that serve as base path for other parameters. If NULL, then, other paths should be absolute paths.

projectConfigurationDirPath

Path to the folder that serve as base path for other paths. If NULL, then, other paths should be absolute paths.

modelFolder

Path to the folder containing pkml simulation files.

paramsFolder

Path to the folder containing excel files with model parameterization;

paramsFile

Name of the excel file with global model parameterization. Must be located in the "paramsFolder".

individualsFile

Name of the excel file with individual-specific model parameterization. Must be located in the "paramsFolder"

populationParamsFile

Name of the excel file with population information. Must be located in the "paramsFolder".

scenarioDefinitionFile

Name of the excel file with scenario definitions. Must be located in the "paramsFolder".

scenarioApplicationsFile

Name of the excel file scenario-specific parameters such as application protocol parameters. Must be located in the "paramsFolder".

plotsFile

Name of the excel file with plot definitions. Must be located in the "paramsFolder".

dataFolder

Path to the folder where experimental data files are located.

dataFile

Name of the excel file with experimental data. Must be located in the "dataFolder"

dataImporterConfigurationFile

Name of data importer configuration file in xml format used to load the data. Must be located in the "dataFolder"

compoundPropertiesFile

Path to the excel file containing information about all compounds in the model. Must be located in the "dataFolder"

outputFolder

Path to the folder where the results should be saved to; relative to the "Code" folder Initialize

Methods

Public methods

ProjectConfiguration$new()

ProjectConfiguration$print()

ProjectConfiguration$save()

ProjectConfiguration$clone()

Method new()

Usage

ProjectConfiguration$new(projectConfigurationFilePath = character())

Arguments

projectConfigurationFilePath

A string representing the path to the project configuration file. Print

Method print()

print prints a summary of the Project Configuration.

Usage

ProjectConfiguration$print()

Method save()

Export ProjectConfiguration object to ProjectConfiguration.xlsx

Usage

ProjectConfiguration$save()

Method clone()

The objects of this class are cloneable with this method.

Usage

ProjectConfiguration$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.