XYData {esqlabsR}	R Documentation

XYData

Description

A set of x and y value pairs

Super classes

ospsuite::Printable -> esqlabsR::Plotable -> XYData

Active bindings

xValues

An array of x-values. For time series, the values must be in minutes.

yValues

An array of y-values.

yError

An array of arithmetic error of the y-values. Only positive values are allowed

xMax

Maximal value of x values plus xOffset multiplied by the scaling factor

xMin

Minimal value of x values plus xOffset multiplied by the scaling factor

yMax

Maximal value (plus error, if specified) of y values plus yOffset multiplied by the scaling factor

yMin

Minimal value (minus error, if specified) of y values plus yOffset multiplied by the scaling factor

dataType

Type of the data. See enum XYDataTypes for the list of supported types.

xDimension

Dimension of x values. See enum ospDimensions for the list of supported dimensions.

xUnit

Unit of x values

yDimension

Dimension of y values. See enum ospDimensions for the list of supported dimensions.

yUnit

Unit of y values

yErrorUnit

Unit of y error values

MW

Molecular weight in g/mol. Required for conversion between molar and mass dimensions. Can be NULL (default)

Methods

Public methods

XYData$new()

XYData$yMinPositive()

XYData$xValuesProcessed()

XYData$yValuesProcessed()

XYData$yErrorProcessed()

XYData$getAllMetaData()

XYData$setMetaData()

XYData$print()

XYData$clone()

Inherited methods

Method new()

Initialize a new instance of the class. xVals, yVals, and yError (optional) must be of the same length

Usage

XYData$new(xVals, yVals, label, yError = NULL)

Arguments

xVals

An array of numeric x values.

yVals

An array of numeric y values,

label

A string that is used as a label (e.g. in the legend) for the data set

yError

An array of numeric values of the arithmetic error. Optional

Returns

A new XYData object.

Method yMinPositive()

Returns the minimal value (minus error, if specified) of the y series that is not negative or null

Usage

XYData$yMinPositive()

Returns

The minimal non-negative value of the y series minus error, if specified, plus yOffset multiplied by the scaling factor If no such value exists, returns Inf.

Method xValuesProcessed()

x values with all conversions applied.

Usage

XYData$xValuesProcessed(unit = NULL)

Arguments

unit

Target unit. If NULL (default), no conversion between units is applied.

Returns

Raw xValues plus xOffset multiplied by xFactor and converted to a specified unit. It is assumed that raw xValues are in xUnit.

Method yValuesProcessed()

y values with all conversions applied.

Usage

XYData$yValuesProcessed(unit = NULL)

Arguments

unit

Target unit. If NULL (default), no conversion between units is applied.

Returns

Raw yValues plus yOffset multiplied by yFactor and converted to a specified unit. It is assumed that raw yValues are in yUnit.

Method yErrorProcessed()

y error values with all conversions applied.

Usage

XYData$yErrorProcessed(unit = NULL)

Arguments

unit

Target unit. If NULL (default), the no conversion between units is applied.

Returns

Raw yError multiplied by yFactor and converted to a specified unit. It is assumed that raw yError are in yUnit. If no error is specified, NULL is returned.

Method getAllMetaData()

Meta data list of XYData object

Usage

XYData$getAllMetaData()

Returns

A named list holding the metadata of this XYData

Method setMetaData()

Adds a new entry to meta data list of XYData object or changes its value if name is already present in meta data. If only name is provided or if value is set to NULL, entry with corresponding name is deleted from meta data.

Usage

XYData$setMetaData(name, value = NULL)

Arguments

name

Name of new meta data list entry

value

Value of new meta data list entry

Method print()

Print the object to the console

Usage

XYData$print(...)

Arguments

...

Rest arguments.

Method clone()

The objects of this class are cloneable with this method.

Usage

XYData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package esqlabsR version 1.2.3 Index]