exportSteadyStateToXLS {esqlabsR}	R Documentation

Export steady-state to excel

Description

Export steady-state to excel

Usage

exportSteadyStateToXLS(
  simulation,
  quantitiesPaths = NULL,
  resultsXLSPath = "",
  steadyStateTime = 1000,
  ignoreIfFormula = TRUE,
  stopIfNotFound = TRUE,
  lowerThreshold = 1e-15
)

Arguments

simulation
Simulation object

quantitiesPaths
List of quantity paths (molecules and/or parameters) for which the steady-state is to be simulated. If NULL (default), all molecules and state variable parameters are considered.

resultsXLSPath
Path to the xls-file where the results will be written to. If the file does not exist, a new file is created. If no path is provided, the file will be created in the same directory where the model file is located.

steadyStateTime
Simulation time (minutes). Must be long enough for system to reach a steady-state. 1000 by default

ignoreIfFormula
If TRUE (default), species and parameters with initial values defined by a formula are not included.

stopIfNotFound
Boolean. If TRUE (default), an error is thrown when results for certain species were not generated. This may happen when due to numerical problems some values cannot be calculated, though the whole simulation converges. Setting this argument to FALSE allows to ignore such errors. Check the outputs for empty values when using this option.

lowerThreshold
Numerical value (in µmol). Any steady-state values below this value are considered as numerical noise and replaced by 0. If lowerThreshold is NULL, no cut-off is applied. Default value is 1e-15.

Details

Simulates a given model to its steady-state and creates an Excel-file with the end values of molecules amounts in all containers and parameter values that have a right-hand-side (state variable parameters).

[Package esqlabsR version 1.2.3 Index]