Carry out and visualize sensitivity analysis (with OSPSuite)

Carry out and visualize sensitivity analysis (with OSPSuite)

Usage

sensitivityCalculation(
  simulation,
  outputPaths,
  parameterPaths,
  variationRange = c(seq(0.1, 1, by = 0.1), seq(2, 10, by = 1)),
  pkParameters = c("C_max", "t_max", "AUC_inf"),
  pkDataFilePath = NULL,
  timeSeriesDataFilePath = NULL
)

Arguments

simulation

An object of type Simulation.

outputPaths

Path (or a vector of paths) to the output(s) for which the sensitivity will be analyzed.

parameterPaths

Path (a single or a vector) of the parameter(s) to be varied.

variationRange

Optional numeric vector defining the scaling of the parameters. The same variation range should be applied to all specified parameters. If not specified, the following vector will be used: c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10).

pkParameters

A vector of names of PK parameters for which the sensitivities should be calculated. For a full set of available PK parameters, run names(ospsuite::StandardPKParameter). By default, only the following parameters will be considered: "C_max", "t_max", "AUC_inf". If NULL, all available PK-parameters will be calculated.

pkDataFilePath, timeSeriesDataFilePath

Paths to spreadsheets in which PK-parameter and time series data should be saved. Default is NULL, meaning the data will not be saved to a spreadsheet.

Note

PK parameter "t_max" in some contexts may not show any variation.

Examples


library(ospsuite)

simPath <- system.file("extdata", "Aciclovir.pkml", package = "esqlabsR")
simulation <- loadSimulation(simPath)
outputPaths <- "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)"
parameterPaths <- c(
  "Aciclovir|Lipophilicity",
  "Applications|IV 250mg 10min|Application_1|ProtocolSchemaItem|Dose",
  "Neighborhoods|Kidney_pls_Kidney_ur|Aciclovir|Glomerular Filtration-GFR|GFR fraction"
)

# extract the results into a list of dataframes
sensitivityCalculation(
  simulation = simulation,
  outputPaths = outputPaths,
  parameterPaths = parameterPaths
)
#> $tsData
#> # A tibble: 27,987 x 11
#>    OutputPath   ParameterFactor ParameterPath ParameterValue  Time Concentration
#>    <chr>                  <dbl> <chr>                  <dbl> <dbl>         <dbl>
#>  1 Organism|Pe~             0.1 Neighborhood~            0.1     0          0   
#>  2 Organism|Pe~             0.1 Neighborhood~            0.1     1          3.26
#>  3 Organism|Pe~             0.1 Neighborhood~            0.1     2          9.14
#>  4 Organism|Pe~             0.1 Neighborhood~            0.1     3         15.1 
#>  5 Organism|Pe~             0.1 Neighborhood~            0.1     4         20.9 
#>  6 Organism|Pe~             0.1 Neighborhood~            0.1     5         26.5 
#>  7 Organism|Pe~             0.1 Neighborhood~            0.1     6         31.8 
#>  8 Organism|Pe~             0.1 Neighborhood~            0.1     7         36.9 
#>  9 Organism|Pe~             0.1 Neighborhood~            0.1     8         41.7 
#> 10 Organism|Pe~             0.1 Neighborhood~            0.1     9         46.2 
#> # ... with 27,977 more rows, and 5 more variables: Unit <chr>, Dimension <chr>,
#> #   TimeUnit <chr>, TimeDimension <chr>, molWeight <dbl>
#> 
#> $pkData
#> # A tibble: 171 x 9
#>    OutputPath           ParameterPath ParameterFactor ParameterValue PKParameter
#>    <chr>                <chr>                   <dbl>          <dbl> <chr>      
#>  1 Organism|Peripheral~ Neighborhood~             0.1            0.1 AUC_inf    
#>  2 Organism|Peripheral~ Neighborhood~             0.2            0.2 AUC_inf    
#>  3 Organism|Peripheral~ Neighborhood~             0.3            0.3 AUC_inf    
#>  4 Organism|Peripheral~ Neighborhood~             0.4            0.4 AUC_inf    
#>  5 Organism|Peripheral~ Neighborhood~             0.5            0.5 AUC_inf    
#>  6 Organism|Peripheral~ Neighborhood~             0.6            0.6 AUC_inf    
#>  7 Organism|Peripheral~ Neighborhood~             0.7            0.7 AUC_inf    
#>  8 Organism|Peripheral~ Neighborhood~             0.8            0.8 AUC_inf    
#>  9 Organism|Peripheral~ Neighborhood~             0.9            0.9 AUC_inf    
#> 10 Organism|Peripheral~ Neighborhood~             1              1   AUC_inf    
#> # ... with 161 more rows, and 4 more variables: PKParameterValue <dbl>,
#> #   Unit <chr>, PercentChangePK <dbl>, SensitivityPKParameter <dbl>
#> 
#> attr(,"class")
#> [1] "SensitivityAnalysis" "list"